requirements.txt
streamlit
tensorflow
numpy
pillow
scikit-learn
openai
plotly
pandas